New metrics for disk usage.
Based on disk-capacity-metrics.rb by bhenerey and nstielau
The difference here being how the key is defined in graphite and the
size we emit to graphite(now using megabytes). Also i dropped inode info.
Using this as an example
Filesystem                                 Size  Used Avail Use% Mounted on
/dev/mapper/precise64-root                  79G  3.5G   72G   5% /
/dev/sda1                                  228M   25M  192M  12% /boot
The keys with this plugin will be
 disk_usage.root and disk_usage.root.boot instead of
 disk.dev.mapper.precise64-root and disk.sda1